Fix parser/network buffer-safety bugs and correctness issues from audit
Security (attacker-controlled RF/BLE/network input):
- CRSF parser: a 0xFE/0xFF length byte wrapped uint8_t expectedLength to 0/1
  and memcpy'd up to ~252 bytes into the 60-byte payload buffer. Range-validate
  the length, widen expectedLength, and read the CRC at the correct offset.
- Remote ID BLE: parsers read fixed offsets (to byte 22) from variable-length
  service data after only a length<2 check. Add per-type minimum-length guards.
- ESP-NOW: recv callback only checked len>=4 before dereferencing union payloads
  at offset 8+. Validate header + per-type payload length before handling.
- Web /api/logs/download: reject path traversal (.., separators, CR/LF).
- Web OTA: reported success and rebooted even when nothing was written; track
  success from the upload and gate on a verified Update.end().
- Web /api/config POST: stop writing one byte past the framework buffer and
  handle multi-chunk bodies via a bounded String; same one-past fix for the WS handler.

Correctness / robustness:
- CC1101 receiveData returned the raw on-air length -> caller OOB read; return
  bytes actually copied. Scans now restore the real pre-scan frequency, and
  writeRegister's busy-wait is bounded with a timeout on PIN_SPI_MISO.
- MAVLink expectedLength widened to uint16_t (payloads 248-255 wrapped);
  parseHeartbeat zero-inits on short payloads.
- signal_database partial-match distance widened from int8_t (overflow).
- power_manager sleepDurationUs widened to uint64_t (>71 min truncation).
- gps addGeofence NUL-terminates the name; RF protocol flags no longer latch
  forever (reset each scan cycle).
- Cosmetic: NRF24 MHz/GHz log, Meshtastic %d for size_t, power_manager comments.

Found via a multi-perspective audit with adversarial verification; only
high-confidence, self-contained fixes applied (deeper concurrency/RX5808/CSV
items left for follow-up as they need hardware verification).

+++ src/meshtastic_client.cpp
@@ -213,7 +213,7 @@ bool MeshtasticClient::triangulateThreat(const char* droneID, double& outLat, do
     }
 
     if (reports.size() < 3) {
-        Serial.printf("[LoRa] Triangulation failed: only %d reports found for %s\n", reports.size(), droneID);
+        Serial.printf("[LoRa] Triangulation failed: only %d reports found for %s\n", (int)reports.size(), droneID);
         return false;
     }
